feat: 카테고리 목록 조회 api 구현 #16

+++ src/main/java/com/umc/yeongkkeul/web/dto/CategoryResponseDTO.java
@@ -27,7 +27,7 @@ public class CategoryResponseDTO {
     @AllArgsConstructor
     public static class CategoryViewListDTO{
         List<CategoryViewDTO> categoryList;
-        Long totalElements;
+        int totalElements;
     }
 
 }
